Updated version. Formatting for phone number implemented, as well as borders and other minor changes.
When running the app, phone verification for the first user works fine. However, if you want to add another user on the simulator, you must sign out, close the simulator and reopen, and then you can add a new user. This is because only one user is intended per device.